requirements.txt
fastapi==0.115.6
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
email-validator==2.2.0
pydantic==2.5.0
